app/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
clikit==0.6.2
comm==0.1.3
convertdate==2.4.0
crashtest==0.3.1
cycler==0.11.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.4
Flask==2.2.5
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
hijri-converter==2.3.1
holidays==0.35
huggingface-hub==0.17.3
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-dash==0.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lightning-utilities==0.8.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
orjson==3.9.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plotly==5.14.1
plotly-resampler==0.8.3.2
prompt-toolkit==3.0.38
prophet==1.1.5
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
pylev==1.4.0
PyMeeus==0.5.12
pyparsing==3.0.9
pysimdjson==5.0.2
pystan==3.7.0
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sklearn-contrib-py-earth @ git+https://github.com/scikit-learn-contrib/py-earth@400f84d435b7277124535c09ca32132c1d0eaa74
smmap==5.0.1
stack-data==0.6.2
stanio==0.3.0
streamlit==1.28.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.9.0
transformers==4.34.1
typing_extensions==4.6.3
tzlocal==5.2
urllib3==1.26.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webargs==8.3.0
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
